Move API to own module
