+++ api/src/main/java/ch/andre601/advancedserverlist/api/exceptions/InvalidPlaceholderProviderException.java
@@ -23,20 +23,15 @@
  *
  */
 
-package ch.andre601.advancedserverlist.api.objects;
+package ch.andre601.advancedserverlist.api.exceptions;
 
 /**
- * Placeholder class used for Platforms that do not allow to retrieve data from specific players that aren't online such
- * as BungeeCord and Velocity.
- * <br>It is also used in the {@link ch.andre601.advancedserverlist.api.internals.placeholders.ServerPlaceholders ServerPlaceholders class}
- * as it doesn't require any player data.
- * 
- * <p>Feel free to use this class in your {@link ch.andre601.advancedserverlist.api.PlaceholderProvider PlaceholderProvider instance}
- * should you not need any player data from the server.
+ * RuntimeException thrown whenever an invalid {@link ch.andre601.advancedserverlist.api.PlaceholderProvider PlaceholderProvider}
+ * has been given.
  */
-public class DummyPlayer extends GenericPlayer<Object>{
-    public DummyPlayer(){
-        this.name = "dummy";
-        this.protocol = 0;
+public class InvalidPlaceholderProviderException extends RuntimeException{
+    
+    public InvalidPlaceholderProviderException(String msg){
+        super(msg);
     }
 }
