+++ api/src/main/java/ch/andre601/advancedserverlist/api/objects/GenericServer.java
@@ -1,0 +1,70 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2022 Andre_601
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ *
+ */
+
+package ch.andre601.advancedserverlist.api.objects;
+
+/**
+ * Simple class used to wrap around some generic server data such as online player count, amount of total players that
+ * can join and the host (IP/Domain) that got pinged by the player.
+ */
+public class GenericServer{
+    
+    private final int playersOnline;
+    private final int playersMax;
+    private final String host;
+    
+    public GenericServer(int playersOnline, int playersMax, String host){
+        this.playersOnline = playersOnline;
+        this.playersMax = playersMax;
+        this.host = host;
+    }
+    
+    /**
+     * Returns the number of players currently online on the server.
+     *
+     * @return Number of players online on the server.
+     */
+    public int getPlayersOnline(){
+        return playersOnline;
+    }
+    
+    /**
+     * Returns the number of total players that can join the server.
+     *
+     * @return Number of total players that can join the server.
+     */
+    public int getPlayersMax(){
+        return playersMax;
+    }
+    
+    /**
+     * Returns the IP/Domain that got pinged by the player.
+     *
+     * @return Possibly-null String containing the IP/Domain that got pinged by the player.
+     */
+    public String getHost(){
+        return host;
+    }
+}
