+++ core/pom.xml
@@ -45,6 +45,11 @@
     
     <dependencies>
         <dependency>
+            <groupId>ch.andre601.advancedserverlist</groupId>
+            <artifactId>api</artifactId>
+            <version>${plugin.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.spongepowered</groupId>
             <artifactId>configurate-yaml</artifactId>
             <version>4.1.2</version>
@@ -79,6 +84,12 @@
             <groupId>com.squareup.okhttp3</groupId>
             <artifactId>okhttp</artifactId>
             <version>4.10.0</version>
+        </dependency>
+        <dependency>
+            <groupId>ch.andre601.advancedserverlist</groupId>
+            <artifactId>parent</artifactId>
+            <version>parent</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
